partial refactoring
