+++ s3-sink-connector/src/main/java/io/aiven/kafka/connect/s3/config/S3SinkConfig.java
@@ -43,17 +43,12 @@ import java.util.Objects;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
+import io.aiven.kafka.connect.common.config.*;
 import org.apache.kafka.common.config.ConfigDef;
 import org.apache.kafka.common.config.ConfigDef.Importance;
 import org.apache.kafka.common.config.ConfigDef.Type;
 import org.apache.kafka.common.config.ConfigException;
 
-import io.aiven.kafka.connect.common.config.CompressionType;
-import io.aiven.kafka.connect.common.config.FixedSetRecommender;
-import io.aiven.kafka.connect.common.config.OutputField;
-import io.aiven.kafka.connect.common.config.OutputFieldEncodingType;
-import io.aiven.kafka.connect.common.config.OutputFieldType;
-import io.aiven.kafka.connect.common.config.TimestampSource;
 import io.aiven.kafka.connect.common.config.validators.FileCompressionTypeValidator;
 import io.aiven.kafka.connect.common.config.validators.FilenameTemplateValidator;
 import io.aiven.kafka.connect.common.config.validators.TimeZoneValidator;
@@ -132,7 +127,7 @@ final public class S3SinkConfig extends S3SinkBaseConfig {
         addAwsConfigGroup(configDef);
         addS3SinkConfig(configDef);
         addAwsStsConfigGroup(configDef);
-        addFileConfigGroup(configDef);
+        FileNameFragment.update(configDef);
         addOutputFieldsFormatConfigGroup(configDef, null);
         addDeprecatedTimestampConfig(configDef);
         addDeprecatedConfiguration(configDef);
@@ -183,61 +178,61 @@ final public class S3SinkConfig extends S3SinkBaseConfig {
         }
     }
 
-    private static void addFileConfigGroup(final ConfigDef configDef) {
-        int fileGroupCounter = 0;
-
-        configDef.define(FILE_NAME_TEMPLATE_CONFIG, ConfigDef.Type.STRING, null,
-                new FilenameTemplateValidator(FILE_NAME_TEMPLATE_CONFIG), ConfigDef.Importance.MEDIUM,
-                "The template for file names on S3. "
-                        + "Supports `{{ variable }}` placeholders for substituting variables. "
-                        + "Currently supported variables are `topic`, `partition`, and `start_offset` "
-                        + "(the offset of the first record in the file). "
-                        + "Only some combinations of variables are valid, which currently are:\n"
-                        + "- `topic`, `partition`, `start_offset`."
-                        + "There is also `key` only variable {{key}} for grouping by keys",
-                GROUP_FILE, fileGroupCounter++, // NOPMD UnusedAssignment
-                ConfigDef.Width.LONG, FILE_NAME_TEMPLATE_CONFIG);
-
-        final String supportedCompressionTypes = CompressionType.names()
-                .stream()
-                .map(f -> "'" + f + "'")
-                .collect(Collectors.joining(", "));
-
-        configDef.define(FILE_COMPRESSION_TYPE_CONFIG, ConfigDef.Type.STRING, null, new FileCompressionTypeValidator(),
-                ConfigDef.Importance.MEDIUM,
-                "The compression type used for files put on S3. " + "The supported values are: "
-                        + supportedCompressionTypes + ".",
-                GROUP_FILE, fileGroupCounter++, // NOPMD UnusedAssignment
-                ConfigDef.Width.NONE, FILE_COMPRESSION_TYPE_CONFIG,
-                FixedSetRecommender.ofSupportedValues(CompressionType.names()));
-
-        configDef.define(FILE_MAX_RECORDS, ConfigDef.Type.INT, 0, new ConfigDef.Validator() {
-            @Override
-            public void ensureValid(final String name, final Object value) {
-                assert value instanceof Integer;
-                if ((Integer) value < 0) {
-                    throw new ConfigException(FILE_MAX_RECORDS, value, "must be a non-negative integer number");
-                }
-            }
-        }, ConfigDef.Importance.MEDIUM,
-                "The maximum number of records to put in a single file. " + "Must be a non-negative integer number. "
-                        + "0 is interpreted as \"unlimited\", which is the default.",
-                GROUP_FILE, fileGroupCounter++, // NOPMD UnusedAssignment
-                ConfigDef.Width.SHORT, FILE_MAX_RECORDS);
-
-        configDef.define(FILE_NAME_TIMESTAMP_TIMEZONE, ConfigDef.Type.STRING, ZoneOffset.UTC.toString(),
-                new TimeZoneValidator(), ConfigDef.Importance.LOW,
-                "Specifies the timezone in which the dates and time for the timestamp variable will be treated. "
-                        + "Use standard shot and long names. Default is UTC",
-                GROUP_FILE, fileGroupCounter++, // NOPMD UnusedAssignment
-                ConfigDef.Width.SHORT, FILE_NAME_TIMESTAMP_TIMEZONE);
-
-        configDef.define(FILE_NAME_TIMESTAMP_SOURCE, ConfigDef.Type.STRING, TimestampSource.Type.WALLCLOCK.name(),
-                new TimestampSourceValidator(), ConfigDef.Importance.LOW,
-                "Specifies the the timestamp variable source. Default is wall-clock.", GROUP_FILE, fileGroupCounter++, // NOPMD
-                                                                                                                       // UnusedAssignment
-                ConfigDef.Width.SHORT, FILE_NAME_TIMESTAMP_SOURCE);
-    }
+//    private static void addFileConfigGroup(final ConfigDef configDef) {
+//        int fileGroupCounter = 0;
+//
+//        configDef.define(FILE_NAME_TEMPLATE_CONFIG, ConfigDef.Type.STRING, null,
+//                new FilenameTemplateValidator(FILE_NAME_TEMPLATE_CONFIG), ConfigDef.Importance.MEDIUM,
+//                "The template for file names on S3. "
+//                        + "Supports `{{ variable }}` placeholders for substituting variables. "
+//                        + "Currently supported variables are `topic`, `partition`, and `start_offset` "
+//                        + "(the offset of the first record in the file). "
+//                        + "Only some combinations of variables are valid, which currently are:\n"
+//                        + "- `topic`, `partition`, `start_offset`."
+//                        + "There is also `key` only variable {{key}} for grouping by keys",
+//                GROUP_FILE, fileGroupCounter++, // NOPMD UnusedAssignment
+//                ConfigDef.Width.LONG, FILE_NAME_TEMPLATE_CONFIG);
+//
+//        final String supportedCompressionTypes = CompressionType.names()
+//                .stream()
+//                .map(f -> "'" + f + "'")
+//                .collect(Collectors.joining(", "));
+//
+//        configDef.define(FILE_COMPRESSION_TYPE_CONFIG, ConfigDef.Type.STRING, null, new FileCompressionTypeValidator(),
+//                ConfigDef.Importance.MEDIUM,
+//                "The compression type used for files put on S3. " + "The supported values are: "
+//                        + supportedCompressionTypes + ".",
+//                GROUP_FILE, fileGroupCounter++, // NOPMD UnusedAssignment
+//                ConfigDef.Width.NONE, FILE_COMPRESSION_TYPE_CONFIG,
+//                FixedSetRecommender.ofSupportedValues(CompressionType.names()));
+//
+//        configDef.define(FILE_MAX_RECORDS, ConfigDef.Type.INT, 0, new ConfigDef.Validator() {
+//            @Override
+//            public void ensureValid(final String name, final Object value) {
+//                assert value instanceof Integer;
+//                if ((Integer) value < 0) {
+//                    throw new ConfigException(FILE_MAX_RECORDS, value, "must be a non-negative integer number");
+//                }
+//            }
+//        }, ConfigDef.Importance.MEDIUM,
+//                "The maximum number of records to put in a single file. " + "Must be a non-negative integer number. "
+//                        + "0 is interpreted as \"unlimited\", which is the default.",
+//                GROUP_FILE, fileGroupCounter++, // NOPMD UnusedAssignment
+//                ConfigDef.Width.SHORT, FILE_MAX_RECORDS);
+//
+//        configDef.define(FILE_NAME_TIMESTAMP_TIMEZONE, ConfigDef.Type.STRING, ZoneOffset.UTC.toString(),
+//                new TimeZoneValidator(), ConfigDef.Importance.LOW,
+//                "Specifies the timezone in which the dates and time for the timestamp variable will be treated. "
+//                        + "Use standard shot and long names. Default is UTC",
+//                GROUP_FILE, fileGroupCounter++, // NOPMD UnusedAssignment
+//                ConfigDef.Width.SHORT, FILE_NAME_TIMESTAMP_TIMEZONE);
+//
+//        configDef.define(FILE_NAME_TIMESTAMP_SOURCE, ConfigDef.Type.STRING, TimestampSource.Type.WALLCLOCK.name(),
+//                new TimestampSourceValidator(), ConfigDef.Importance.LOW,
+//                "Specifies the the timestamp variable source. Default is wall-clock.", GROUP_FILE, fileGroupCounter++, // NOPMD
+//                                                                                                                       // UnusedAssignment
+//                ConfigDef.Width.SHORT, FILE_NAME_TIMESTAMP_SOURCE);
+//    }
 
     private static void addDeprecatedTimestampConfig(final ConfigDef configDef) {
         int timestampGroupCounter = 0;
@@ -297,14 +292,12 @@ final public class S3SinkConfig extends S3SinkBaseConfig {
 
     @Override
     public List<OutputField> getOutputFields() {
-        if (Objects.nonNull(getList(FORMAT_OUTPUT_FIELDS_CONFIG))
-                && !get(FORMAT_OUTPUT_FIELDS_CONFIG).equals(ConfigDef.NO_DEFAULT_VALUE)) {
-            return getOutputFields(FORMAT_OUTPUT_FIELDS_CONFIG);
+        List<OutputField> result = super.getOutputFields();
+
+        if (result == null) {
+            result = new OutputFormatFragment(this).getOutputFields(OUTPUT_FIELDS);
         }
-        if (Objects.nonNull(getList(OUTPUT_FIELDS)) && !get(OUTPUT_FIELDS).equals(ConfigDef.NO_DEFAULT_VALUE)) {
-            return getOutputFields(OUTPUT_FIELDS);
-        }
-        return List.of(new OutputField(OutputFieldType.VALUE, OutputFieldEncodingType.BASE64));
+        return result != null ? result :  List.of(new OutputField(OutputFieldType.VALUE, OutputFieldEncodingType.BASE64));
     }
 
     public List<OutputField> getOutputFields(final String format) {
@@ -317,12 +310,12 @@ final public class S3SinkConfig extends S3SinkBaseConfig {
         }).collect(Collectors.toUnmodifiableList());
     }
 
-    @Override
-    public OutputFieldEncodingType getOutputFieldEncodingType() {
-        return Objects.nonNull(getString(FORMAT_OUTPUT_FIELDS_VALUE_ENCODING_CONFIG))
-                ? OutputFieldEncodingType.forName(getString(FORMAT_OUTPUT_FIELDS_VALUE_ENCODING_CONFIG))
-                : OutputFieldEncodingType.BASE64;
-    }
+//    @Override
+//    public OutputFieldEncodingType getOutputFieldEncodingType() {
+//        return Objects.nonNull(getString(FORMAT_OUTPUT_FIELDS_VALUE_ENCODING_CONFIG))
+//                ? OutputFieldEncodingType.forName(getString(FORMAT_OUTPUT_FIELDS_VALUE_ENCODING_CONFIG))
+//                : OutputFieldEncodingType.BASE64;
+//    }
 
     public Template getPrefixTemplate() {
         final var template = Template.of(getAwsS3Prefix());
